+++ backend-src/UZonMailService/Models/SQL/Emails/Outbox.cs
@@ -45,7 +45,7 @@ namespace UZonMailService.Models.SQL.Emails
         /// <summary>
         /// 代理 Id
         /// </summary>
-        public int ProxyId { get; set; }
+        public long ProxyId { get; set; }
 
         /// <summary>
         /// 单日最大发送数量
@@ -70,7 +70,7 @@ namespace UZonMailService.Models.SQL.Emails
         /// <param name="userSetting"></param>
         /// <param name="groupId"></param>
         /// <returns></returns>
-        public OutboxEmailAddress ToOutboxEmailAddress(UserSetting userSetting, int groupId, List<string> smtpPasswordSecretKeys)
+        public OutboxEmailAddress ToOutboxEmailAddress(UserSetting userSetting, long groupId, List<string> smtpPasswordSecretKeys)
         {
             // 对密码进行解密
             var plainPassword = Password.DeAES(smtpPasswordSecretKeys[0], smtpPasswordSecretKeys[1]);
